Major overhaul of flow structure - Doubles converted to BigDecimals
For numerical accuracy, all calculations for flow shifts (and several
other places) are now done using BigDecimal objects. This includes the
maxDelta and deltaH calculations in equilibrateBush, alterBushFlow
calculations, Dijkstra & topoSearch algorithms, longest & shortest path
cost calculations, improveBush comparisons, ancillary calculations, and
several link characteristic calculations.

Removed getBushFlow method in Bush. A short-circuit is now provided in
equilibrateBush if the longest path has no flow on it. The main class is
reorganized to include settings at top of the file.

+++ src/edu/utexas/wrap/Network.java
@@ -176,7 +176,7 @@ public class Network {
 		Double tstt = 0.0;
 		
 		for(Link l: getLinks()){
-			tstt += l.getFlow() * l.getTravelTime();
+			tstt += l.getFlow().doubleValue() * l.getTravelTime().doubleValue();
 		}
 		return tstt;
 	}
@@ -188,7 +188,7 @@ public class Network {
 		for (Link l : getLinks()) {
 			for (Origin o : origins) {
 				for (Bush b : o.getBushes()) {
-					numerator += l.getBushFlow(b) * l.getPrice(b.getVOT());
+					numerator += l.getBushFlow(b).doubleValue() * l.getPrice(b.getVOT()).doubleValue();
 				}
 			}
 		}
@@ -200,7 +200,7 @@ public class Network {
 					
 					Double demand = b.getDemand(d.getID());
 					try {
-						denominator += b.getL(d) * demand;
+						denominator += b.getL(d).doubleValue() * demand;
 					} catch (UnreachableException e) {
 						if (e.demand > 0.0)	e.printStackTrace();
 					}
@@ -234,7 +234,7 @@ public class Network {
 	public Double Beckmann() {
 		Double b = 0.0;
 		for (Link l : getLinks()) {
-			b += l.tIntegral();
+			b += l.tIntegral().doubleValue();
 		}
 		return b;
 	}
@@ -259,7 +259,7 @@ public class Network {
 			Double sum = 0.0;
 			for (Origin o : origins) {
 				for (Bush b : o.getBushes()) {
-						sum += l.getBushFlow(b);	
+						sum += l.getBushFlow(b).doubleValue();	
 				}
 			}
 			out.println(l+"\t"+sum);
